+++ NEWS.md
@@ -1,9 +1,9 @@
 # camtrapR devel
 
 ## Other changes
-* recordTable & filterRecordTable now output an object of class ctrpr_rectbl with dedicated print method
-* cameraOperation now outputs an object of class ctrpr_opermat with dedicated print and plot methods
-* detectionHistory now outputs an object of class ctrpr_dethist with a dedicated print method
+* recordTable & filterRecordTable now output an object of class records with dedicated print method
+* cameraOperation now outputs an object of class camOp with dedicated print and plot methods
+* detectionHistory now outputs an object of class detHist with a dedicated print method
 
 
 # camtrapR 3.0.6